+++ src/recordTest/java/com/surrealdb/pojos/EmailRecord.java
@@ -1,0 +1,4 @@
+package com.surrealdb.pojos;
+
+public record EmailRecord(String address, NameRecord name) {
+}