+++ src/recordTest/java/com/surrealdb/pojos/NameRecord.java
@@ -1,0 +1,4 @@
+package com.surrealdb.pojos;
+
+public record NameRecord(String first, String last) {
+}